+++ src/Components/Store.lua
@@ -2,7 +2,8 @@ local BasicState = require(script.Parent.Parent.Libs.BasicState)
 
 local State = BasicState.new({
   Material = Enum.Material.Grass,
-  Panel = 'Materials'
+  Panel = 'Materials',
+  SearchTerm = ''
 })
 
 State.ProtectType = true